Refactor views and templates (Customer => User)
Customer model has been replaced by built-in User
