+++ lupanes/templates/lupanes/base.html
@@ -80,15 +80,15 @@
             <li class="nav-item dropdown">
               <!-- <button type="button" class="btn text-light rounded-circle btn-dark"
                 data-bs-toggle="dropdown" aria-expanded="false" data-bs-display="static">
-                {{ customer.name|slice:":1" }}
+                {{ customer.username|slice:":1" }}
               </button> -->
               <button type="button" class="btn btn-link nav-link py-2 px-0 px-lg-2 dropdown-toggle"
                 data-bs-toggle="dropdown" aria-expanded="false" data-bs-display="static">
-                {{ customer.name }}
+                {{ request.user.username }}
               </button>
               <ul class="dropdown-menu dropdown-menu-end">
                 <li>
-                  <a class="dropdown-item" href="{% url 'lupanes:customer-logout' %}">
+                  <a class="dropdown-item" href="{% url 'users:logout' %}">
                     salir
                   </a>
                 </li>
