+++ lupanes/users/templates/users/login.html
@@ -16,9 +16,6 @@
     </form>
 </div>
 <div class="row-fluid mt-3 pt-3 border-top border-dark text-center">
-  <p>¿No tienes nombre de usuario? <br>
-    <i class="fa-solid fa-arrow-down"></i></p>
-  <a class="btn btn-success form-control" href="{% url 'lupanes:customer-login' %}">
-    Accede con tu nombre de nevera</a>
+  <p>Tu nombre de usuario es el nombre de tu nevera</p>
 </div><!-- /.row-fluid -->
 {% endblock login_content %}
