+++ lupanes/views/__init__.py
@@ -1,6 +1,4 @@
 from lupanes.views.customer import (
-    CustomerLoginView,
-    CustomerLogoutView,
     DeliveryNoteCreateView,
     DeliveryNoteDeleteView,
     DeliveryNoteUpdateView,
@@ -12,8 +10,6 @@ from lupanes.views.customer import (
 from lupanes.views.manager import DeliveryNoteListView, DeliveryNoteSummaryView
 
 __all__ = [
-    "CustomerLoginView",
-    "CustomerLogoutView",
     "DeliveryNoteCreateView",
     "DeliveryNoteDeleteView",
     "DeliveryNoteUpdateView",
