+++ proj/urls.py
@@ -20,5 +20,5 @@ from django.views.generic import RedirectView
 urlpatterns = [
     path('', include('lupanes.urls')),
     path('', include('lupanes.users.urls')),
-    path('', RedirectView.as_view(pattern_name='lupanes:customer-login', permanent=True), name='root_index'),
+    path('', RedirectView.as_view(pattern_name='users:login', permanent=True), name='root_index'),
 ]
